+++ snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/datastore/request/SnomedExtendedMapMemberCreateDelegate.java
@@ -49,15 +49,14 @@ final class SnomedExtendedMapMemberCreateDelegate extends SnomedRefSetMemberCrea
 		checkComponentExists(refSet, context, SnomedRf2Headers.FIELD_CORRELATION_ID);
 		checkComponentExists(refSet, context, SnomedRf2Headers.FIELD_MAP_CATEGORY_ID);
 
-		// FIXME: narrowing conversion from integer to byte for group and priority
 		SnomedComplexMapRefSetMember member = SnomedComponents.newComplexMapMember()
 				.withActive(isActive())
 				.withReferencedComponent(getReferencedComponentId())
 				.withModule(getModuleId())
 				.withRefSet(getReferenceSetId())
 				.withMapTargetId(getComponentId(SnomedRf2Headers.FIELD_MAP_TARGET))
-				.withGroup(getProperty(SnomedRf2Headers.FIELD_MAP_GROUP, Integer.class).byteValue())
-				.withPriority(getProperty(SnomedRf2Headers.FIELD_MAP_PRIORITY, Integer.class).byteValue())
+				.withGroup(getProperty(SnomedRf2Headers.FIELD_MAP_GROUP, Integer.class))
+				.withPriority(getProperty(SnomedRf2Headers.FIELD_MAP_PRIORITY, Integer.class))
 				.withMapRule(getProperty(SnomedRf2Headers.FIELD_MAP_RULE))
 				.withMapAdvice(getProperty(SnomedRf2Headers.FIELD_MAP_ADVICE))
 				.withCorrelationId(getComponentId(SnomedRf2Headers.FIELD_CORRELATION_ID))
