+++ snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/datastore/request/SnomedExtendedMapMemberUpdateDelegate.java
@@ -47,13 +47,13 @@ final class SnomedExtendedMapMemberUpdateDelegate extends SnomedRefSetMemberUpda
 			changed |= true;
 		}
 
-		if (newGroup != null && newGroup.byteValue() != concreteDomainMember.getMapGroup()) {
-			concreteDomainMember.setMapGroup(newGroup.byteValue());
+		if (newGroup != null && newGroup.intValue() != concreteDomainMember.getMapGroup()) {
+			concreteDomainMember.setMapGroup(newGroup);
 			changed |= true;
 		}
 
-		if (newPriority != null && newPriority.byteValue() != concreteDomainMember.getMapPriority()) {
-			concreteDomainMember.setMapPriority(newPriority.byteValue());
+		if (newPriority != null && newPriority.intValue() != concreteDomainMember.getMapPriority()) {
+			concreteDomainMember.setMapPriority(newPriority);
 			changed |= true;
 		}
